refactor: capitalize, translatedDifficulty 리팩토링
-다른 페이지에서도 사용될 translatedDifficulty함수 utils로 리팩토링
