+++ src/utils/translatedDifficulty.ts
@@ -1,0 +1,12 @@
+const translatedDifficulty = (difficulty: string) => {
+  switch (difficulty) {
+    case "easy":
+      return "하";
+    case "medium":
+      return "중";
+    case "hard":
+      return "상";
+  }
+};
+
+export default translatedDifficulty;
